+++ apps/nowait-admin/src/utils/AdminApi.tsx
@@ -11,7 +11,7 @@ const AdminApi = axios.create({
 
 AdminApi.interceptors.request.use(
   (config) => {
-    const token = localStorage.getItem("accessToken");
+    const token = localStorage.getItem("adminToken");
     console.log(token, "토큰 알려줘");
     if (token) {
       config.headers.Authorization = `Bearer ${token}`;
